+++ src/util.cpp
@@ -98,7 +98,7 @@ bool fLogIPs = DEFAULT_LOGIPS;
 std::atomic<bool> fReopenDebugLog(false);
 CTranslationInterface translationInterface;
 
-int currentAlgo = 0;
+uint8_t currentAlgo = 0;
 
 /** Log categories bitfield. */
 std::atomic<uint32_t> logCategories(0);
